Move crate::locations to crate::location in SDK
Module names for Grid features should generally be singular; in this
instance, we talk about "Grid Location" not "Grid Locations".
