+++ contracts/location/src/state.rs
@@ -23,7 +23,7 @@ cfg_if! {
 }
 
 use grid_sdk::{
-    locations::addressing::compute_gs1_location_address,
+    location::addressing::compute_gs1_location_address,
     pike::addressing::compute_organization_address,
     protocol::{
         location::state::{Location, LocationList, LocationListBuilder},
