+++ daemon/src/event/error.rs
@@ -19,7 +19,7 @@ use std::error::Error;
 use std::fmt;
 
 use grid_sdk::{
-    commits::store::CommitStoreError, locations::store::LocationStoreError,
+    commits::store::CommitStoreError, location::store::LocationStoreError,
     pike::store::PikeStoreError, products::store::ProductStoreError,
     schemas::store::SchemaStoreError, track_and_trace::store::TrackAndTraceStoreError,
 };
