+++ sdk/src/location/store/diesel/operations/update_location.rs
@@ -13,15 +13,15 @@
 // limitations under the License.
 
 use super::LocationStoreOperations;
-use crate::locations::store::diesel::{
+use crate::location::store::diesel::{
     schema::{location, location_attribute},
     LocationStoreError,
 };
 
 use crate::commits::MAX_COMMIT_NUM;
 use crate::error::{ConstraintViolationError, ConstraintViolationType, InternalError};
-use crate::locations::store::diesel::models::{
-    LocationAttributeModel, LocationModel, NewLocationAttributeModel, NewLocationModel,
+use crate::location::store::diesel::models::{
+    LocationModel, NewLocationAttributeModel, NewLocationModel,
 };
 use diesel::{
     dsl::{insert_into, update},
@@ -29,8 +29,8 @@ use diesel::{
     result::{DatabaseErrorKind, Error as dsl_error},
 };
 
-pub(in crate::locations::store::diesel) trait LocationStoreAddLocationOperation {
-    fn add_location(
+pub(in crate::location::store::diesel) trait LocationStoreUpdateLocationOperation {
+    fn update_location(
         &self,
         location: NewLocationModel,
         attributes: Vec<NewLocationAttributeModel>,
@@ -39,22 +39,21 @@ pub(in crate::locations::store::diesel) trait LocationStoreAddLocationOperation 
 }
 
 #[cfg(feature = "postgres")]
-impl<'a> LocationStoreAddLocationOperation
+impl<'a> LocationStoreUpdateLocationOperation
     for LocationStoreOperations<'a, diesel::pg::PgConnection>
 {
-    fn add_location(
+    fn update_location(
         &self,
         location: NewLocationModel,
         attributes: Vec<NewLocationAttributeModel>,
         current_commit_num: i64,
     ) -> Result<(), LocationStoreError> {
         self.conn.transaction::<_, LocationStoreError, _>(|| {
-            let duplicate_loc = location::table
+            let loc = location::table
                 .filter(
                     location::location_id
                         .eq(&location.location_id)
-                        .and(location::service_id.eq(&location.service_id))
-                        .and(location.end_commit_num.eq(&MAX_COMMIT_NUM)),
+                        .and(location::service_id.eq(&location.service_id)),
                 )
                 .first::<LocationModel>(self.conn)
                 .map(Some)
@@ -69,7 +68,7 @@ impl<'a> LocationStoreAddLocationOperation
                     LocationStoreError::InternalError(InternalError::from_source(Box::new(err)))
                 })?;
 
-            if duplicate_loc.is_some() {
+            if loc.is_some() {
                 update(location::table)
                     .filter(
                         location::location_id
@@ -129,65 +128,41 @@ impl<'a> LocationStoreAddLocationOperation
                     }
                 })?;
 
-            for attr in attributes {
-                let duplicate_attr = location_attribute::table
-                    .filter(
-                        location_attribute::location_id
-                            .eq(&attr.location_id)
-                            .and(location_attribute::property_name.eq(&attr.property_name))
-                            .and(location_attribute::service_id.eq(&attr.service_id))
-                            .and(location_attribute::end_commit_num.eq(MAX_COMMIT_NUM)),
-                    )
-                    .first::<LocationAttributeModel>(self.conn)
-                    .map(Some)
-                    .or_else(|err| {
-                        if err == dsl_error::NotFound {
-                            Ok(None)
-                        } else {
-                            Err(err)
-                        }
-                    })
-                    .map_err(|err| {
-                        LocationStoreError::InternalError(InternalError::from_source(Box::new(err)))
-                    })?;
-
-                if duplicate_attr.is_some() {
-                    update(location_attribute::table)
-                        .filter(
-                            location_attribute::location_id
-                                .eq(&attr.location_id)
-                                .and(location_attribute::property_name.eq(&attr.property_name))
-                                .and(location_attribute::service_id.eq(&attr.service_id))
-                                .and(location_attribute::end_commit_num.eq(MAX_COMMIT_NUM)),
-                        )
-                        .set(location_attribute::end_commit_num.eq(current_commit_num))
-                        .execute(self.conn)
-                        .map(|_| ())
-                        .map_err(|err| match err {
-                            dsl_error::DatabaseError(DatabaseErrorKind::UniqueViolation, _) => {
-                                LocationStoreError::ConstraintViolationError(
-                                    ConstraintViolationError::from_source_with_violation_type(
-                                        ConstraintViolationType::Unique,
-                                        Box::new(err),
-                                    ),
-                                )
-                            }
-                            dsl_error::DatabaseError(DatabaseErrorKind::ForeignKeyViolation, _) => {
-                                LocationStoreError::ConstraintViolationError(
-                                    ConstraintViolationError::from_source_with_violation_type(
-                                        ConstraintViolationType::ForeignKey,
-                                        Box::new(err),
-                                    ),
-                                )
-                            }
-                            _ => LocationStoreError::InternalError(InternalError::from_source(
+            update(location_attribute::table)
+                .filter(
+                    location_attribute::location_id
+                        .eq(&location.location_id)
+                        .and(location_attribute::service_id.eq(&location.service_id))
+                        .and(location_attribute::end_commit_num.eq(MAX_COMMIT_NUM)),
+                )
+                .set(location_attribute::end_commit_num.eq(current_commit_num))
+                .execute(self.conn)
+                .map(|_| ())
+                .map_err(|err| match err {
+                    dsl_error::DatabaseError(DatabaseErrorKind::UniqueViolation, _) => {
+                        LocationStoreError::ConstraintViolationError(
+                            ConstraintViolationError::from_source_with_violation_type(
+                                ConstraintViolationType::Unique,
                                 Box::new(err),
-                            )),
-                        })?;
-                }
+                            ),
+                        )
+                    }
+                    dsl_error::DatabaseError(DatabaseErrorKind::ForeignKeyViolation, _) => {
+                        LocationStoreError::ConstraintViolationError(
+                            ConstraintViolationError::from_source_with_violation_type(
+                                ConstraintViolationType::ForeignKey,
+                                Box::new(err),
+                            ),
+                        )
+                    }
+                    _ => {
+                        LocationStoreError::InternalError(InternalError::from_source(Box::new(err)))
+                    }
+                })?;
 
+            for attribute in attributes {
                 insert_into(location_attribute::table)
-                    .values(&attr)
+                    .values(&attribute)
                     .execute(self.conn)
                     .map(|_| ())
                     .map_err(|err| match err {
@@ -219,22 +194,21 @@ impl<'a> LocationStoreAddLocationOperation
 }
 
 #[cfg(feature = "sqlite")]
-impl<'a> LocationStoreAddLocationOperation
+impl<'a> LocationStoreUpdateLocationOperation
     for LocationStoreOperations<'a, diesel::sqlite::SqliteConnection>
 {
-    fn add_location(
+    fn update_location(
         &self,
         location: NewLocationModel,
         attributes: Vec<NewLocationAttributeModel>,
         current_commit_num: i64,
     ) -> Result<(), LocationStoreError> {
         self.conn.transaction::<_, LocationStoreError, _>(|| {
-            let duplicate_loc = location::table
+            let loc = location::table
                 .filter(
                     location::location_id
                         .eq(&location.location_id)
-                        .and(location::service_id.eq(&location.service_id))
-                        .and(location.end_commit_num.eq(&MAX_COMMIT_NUM)),
+                        .and(location::service_id.eq(&location.service_id)),
                 )
                 .first::<LocationModel>(self.conn)
                 .map(Some)
@@ -249,7 +223,7 @@ impl<'a> LocationStoreAddLocationOperation
                     LocationStoreError::InternalError(InternalError::from_source(Box::new(err)))
                 })?;
 
-            if duplicate_loc.is_some() {
+            if loc.is_some() {
                 update(location::table)
                     .filter(
                         location::location_id
@@ -309,65 +283,41 @@ impl<'a> LocationStoreAddLocationOperation
                     }
                 })?;
 
-            for attr in attributes {
-                let duplicate_attr = location_attribute::table
-                    .filter(
-                        location_attribute::location_id
-                            .eq(&attr.location_id)
-                            .and(location_attribute::property_name.eq(&attr.property_name))
-                            .and(location_attribute::service_id.eq(&attr.service_id))
-                            .and(location_attribute::end_commit_num.eq(MAX_COMMIT_NUM)),
-                    )
-                    .first::<LocationAttributeModel>(self.conn)
-                    .map(Some)
-                    .or_else(|err| {
-                        if err == dsl_error::NotFound {
-                            Ok(None)
-                        } else {
-                            Err(err)
-                        }
-                    })
-                    .map_err(|err| {
-                        LocationStoreError::InternalError(InternalError::from_source(Box::new(err)))
-                    })?;
-
-                if duplicate_attr.is_some() {
-                    update(location_attribute::table)
-                        .filter(
-                            location_attribute::location_id
-                                .eq(&attr.location_id)
-                                .and(location_attribute::property_name.eq(&attr.property_name))
-                                .and(location_attribute::service_id.eq(&attr.service_id))
-                                .and(location_attribute::end_commit_num.eq(MAX_COMMIT_NUM)),
-                        )
-                        .set(location_attribute::end_commit_num.eq(current_commit_num))
-                        .execute(self.conn)
-                        .map(|_| ())
-                        .map_err(|err| match err {
-                            dsl_error::DatabaseError(DatabaseErrorKind::UniqueViolation, _) => {
-                                LocationStoreError::ConstraintViolationError(
-                                    ConstraintViolationError::from_source_with_violation_type(
-                                        ConstraintViolationType::Unique,
-                                        Box::new(err),
-                                    ),
-                                )
-                            }
-                            dsl_error::DatabaseError(DatabaseErrorKind::ForeignKeyViolation, _) => {
-                                LocationStoreError::ConstraintViolationError(
-                                    ConstraintViolationError::from_source_with_violation_type(
-                                        ConstraintViolationType::ForeignKey,
-                                        Box::new(err),
-                                    ),
-                                )
-                            }
-                            _ => LocationStoreError::InternalError(InternalError::from_source(
+            update(location_attribute::table)
+                .filter(
+                    location_attribute::location_id
+                        .eq(&location.location_id)
+                        .and(location_attribute::service_id.eq(&location.service_id))
+                        .and(location_attribute::end_commit_num.eq(MAX_COMMIT_NUM)),
+                )
+                .set(location_attribute::end_commit_num.eq(current_commit_num))
+                .execute(self.conn)
+                .map(|_| ())
+                .map_err(|err| match err {
+                    dsl_error::DatabaseError(DatabaseErrorKind::UniqueViolation, _) => {
+                        LocationStoreError::ConstraintViolationError(
+                            ConstraintViolationError::from_source_with_violation_type(
+                                ConstraintViolationType::Unique,
                                 Box::new(err),
-                            )),
-                        })?;
-                }
+                            ),
+                        )
+                    }
+                    dsl_error::DatabaseError(DatabaseErrorKind::ForeignKeyViolation, _) => {
+                        LocationStoreError::ConstraintViolationError(
+                            ConstraintViolationError::from_source_with_violation_type(
+                                ConstraintViolationType::ForeignKey,
+                                Box::new(err),
+                            ),
+                        )
+                    }
+                    _ => {
+                        LocationStoreError::InternalError(InternalError::from_source(Box::new(err)))
+                    }
+                })?;
 
+            for attribute in attributes {
                 insert_into(location_attribute::table)
-                    .values(&attr)
+                    .values(&attribute)
                     .execute(self.conn)
                     .map(|_| ())
                     .map_err(|err| match err {
