+++ sdk/src/location/store/error.rs
