+++ sdk/src/location/store/mod.rs
