+++ sdk/src/store/sqlite.rs
@@ -41,10 +41,8 @@ impl StoreFactory for SqliteStoreFactory {
     }
 
     #[cfg(all(feature = "diesel", feature = "location-store-sqlite"))]
-    fn get_grid_location_store(&self) -> Box<dyn crate::locations::LocationStore> {
-        Box::new(crate::locations::DieselLocationStore::new(
-            self.pool.clone(),
-        ))
+    fn get_grid_location_store(&self) -> Box<dyn crate::location::LocationStore> {
+        Box::new(crate::location::DieselLocationStore::new(self.pool.clone()))
     }
 
     #[cfg(all(feature = "diesel", feature = "product-store-sqlite"))]
